Ajuste - Inserindo pasta usecase e ajustando os testes
