+++ app/src/main/java/br/com/appfastfood/configuracoes/BeanConfiguration.java
@@ -10,8 +10,8 @@ import br.com.appfastfood.pedido.dominio.repositorios.PedidoRepositorio;
 import br.com.appfastfood.pedido.usecase.adaptadores.PedidoServicoImpl;
 import br.com.appfastfood.pedido.usecase.portas.PedidoServico;
 import br.com.appfastfood.produto.dominio.repositorios.ProdutoRepositorio;
-import br.com.appfastfood.produto.dominio.servicos.adaptadores.ProdutoServicoImpl;
-import br.com.appfastfood.produto.dominio.servicos.portas.ProdutoServico;
+import br.com.appfastfood.produto.usecase.adaptadores.ProdutoServicoImpl;
+import br.com.appfastfood.produto.usecase.portas.ProdutoServico;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
