+++ app/src/test/java/br/com/appfastfood/produto/dominio/vo/CategoriaTest.java
@@ -1,5 +1,6 @@
 package br.com.appfastfood.produto.dominio.vo;
 
+import br.com.appfastfood.configuracoes.execption.BadRequestException;
 import br.com.appfastfood.produto.dominio.vo.enums.CategoriaEnum;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +12,7 @@ class CategoriaTest {
     @Test
     void construtor_DeveLancarExcecao_QuandoCategoriaNaoExistir() {
         String categoriaNaoExistente = "CategoriaInexistente";
-        assertThrows(CategoriaNaoEncontradaException.class, () -> new Categoria(categoriaNaoExistente));
+        assertThrows(BadRequestException.class, () -> new Categoria(categoriaNaoExistente));
     }
 
     @Test
